add tests for UMD modules that reference closure variables
As mentioned in #33, popular libs often do their main work outside of
the factory. Thus the factory is reduced to return a variable that is
declared and defined in the enclosing closure. UMD boilerplate is
usually an afterthought so to preserve the code outside of the factory
is very important!