Preserve code outside of factory function.
The AST is no longer replaced in its entirety, instead only the define
call is replaced. This ensures that any code outside of the factory but
still within its scope is preserved. We only wish to transform the
define call, not remove what we *think* is superfluous.

Also the closure of the factory is preserved. If the factory is a
function then it should remain as a function call incase variable names
are shared with its enclosing clojure (as is the case in
umd-webamd-with-function test).

Fixes #33